refactor(ui): move tailwind variant maps from lib into presentation
The button/form/input/section variant maps are the app's own design-system
tokens, consumed only by presentation components, so they belong in the
presentation layer rather than lib. Move them to common/presentation/ui/variants
(named for the *Variants exports) and rewrite consumers onto @/ui/variants/*.
