+++ src/common/presentation/ui/form/compounds/select/select.tsx
@@ -6,8 +6,8 @@ import type {
 } from 'react-hook-form';
 import { twMerge } from 'tailwind-merge';
 
-import type { InputVariants } from '@/lib/tailwindcss/input';
-import { inputVariants } from '@/lib/tailwindcss/input';
+import type { InputVariants } from '@/ui/variants/input';
+import { inputVariants } from '@/ui/variants/input';
 
 import { useForm } from '../../form';
 import { InputErrorText } from '../input/error-text/error-text';
