+++ src/common/presentation/ui/icon-button/icon-button.tsx
@@ -1,8 +1,8 @@
 import type { ComponentProps, MouseEventHandler, ReactNode } from 'react';
 import { twMerge } from 'tailwind-merge';
 
-import type { ButtonVariants } from '@/lib/tailwindcss/button';
-import { buttonVariants } from '@/lib/tailwindcss/button';
+import type { ButtonVariants } from '@/ui/variants/button';
+import { buttonVariants } from '@/ui/variants/button';
 
 export type IconButtonProps = ComponentProps<'button'> & {
   children: ReactNode;
