+++ src/common/presentation/ui/variants/section.ts
