+++ src/module/dashboard/ui/section/section.tsx
@@ -1,8 +1,8 @@
 import type { ComponentProps } from 'react';
 import { twMerge } from 'tailwind-merge';
 
-import type { SectionVariants } from '@/lib/tailwindcss/section';
-import { sectionVariants } from '@/lib/tailwindcss/section';
+import type { SectionVariants } from '@/ui/variants/section';
+import { sectionVariants } from '@/ui/variants/section';
 
 import { SectionControls } from './compounds/controls/controls';
 import { SectionHeading } from './compounds/heading/heading';
